+++ pojo/src/main/java/com/example/dto/BorrowDTO.java
@@ -9,12 +9,12 @@ import java.util.List;
 @Data
 public class BorrowDTO {
 
-    private Long id;
+    private Integer id;
 
     // 借阅状态,有未归还,已归还,还有丢失这些
     private Integer status;
     //借阅用户id
-    private Long userId;
+    private Integer userId;
     //借阅时间
     private LocalDateTime startTime;
     //截止日期
