+++ pojo/src/main/java/com/example/entity/Borrow.java
@@ -13,13 +13,13 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Borrow implements Serializable {
-    private Long id;
+    private Integer id;
     // 借阅状态,有未归还,已归还,还有丢失这些
     private Integer status;
     //借阅用户id
-    private Long userId;
+    private Integer userId;
     //借阅的书籍id
-    private Long bookId;
+    private Integer bookId;
     //借阅时间
     private LocalDateTime startTime;
     //截止日期
