+++ pojo/src/main/java/com/example/entity/Category.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class Category implements Serializable {
 
-    private Long id;
+    private Integer id;
     //分类名称
     private String name;
 
@@ -25,8 +25,8 @@ public class Category implements Serializable {
     private LocalDateTime updateTime;
 
     //创建人
-    private Long createUser;
+    private Integer createUser;
 
     //修改人
-    private Long updateUser;
+    private Integer updateUser;
 }
