+++ pojo/src/main/java/com/example/entity/Employee.java
@@ -32,7 +32,7 @@ public class Employee {
     //@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime updateTime;
 
-    private Long createUser;
+    private Integer createUser;
 
-    private Long updateUser;
+    private Integer updateUser;
 }
