+++ server/src/main/java/com/example/Interceptor/JwtTokenEmployeeInterceptor.java
@@ -48,7 +48,7 @@ public class JwtTokenEmployeeInterceptor implements HandlerInterceptor {
             log.info("jwt校验:{}", token);
             Claims claims = JwtUtil.parseJWT(jwtProperties.getEmployeeSecretKey(), token);
             //校验成功后解析JWT令牌里面的ID,转化为long类型的
-            Long id = Long.valueOf(claims.get(JwtClaimsConstant.LOGIN_ID).toString());
+            Integer id = Integer.valueOf(claims.get(JwtClaimsConstant.LOGIN_ID).toString());
             //然后借助ThreadLocal来存放这个empId到这个线程的空间中去,后续Service层中对Employee对象赋值时再取出来
             BaseContext.setCurrentId(id);
